fix: add credentials to all API requests to send auth cookie

+++ backend/router/task.js
@@ -6,7 +6,7 @@ const router = express.Router();
 
 router.get("/", async (req, res) => {
     try {
-        const tasks = await Task.find();
+        const tasks = await Task.find({createdBy:req.user._id});
         res.json({ success: true, tasks });
     } catch (err) {
         res.status(500).json({ success: false, message: err.message });
@@ -18,7 +18,9 @@ router.post("/add", async (req, res) => {
     try {
         const { title, desc } = req.body;
         const taskData = await Task.create({
-            title, desc
+            title, 
+            desc,
+            createdBy:req.user._id,
         });
         res.json({ success: true, taskData });
     } catch (err){
@@ -29,7 +31,7 @@ router.post("/add", async (req, res) => {
 
 router.delete("/delete/:id", async (req, res) => {
     try{
-        const taskitem = await Task.findByIdAndDelete(req.params.id);
+        const taskitem = await Task.findOneAndDelete({_id:req.params.id, createdBy:req.user._id});
         res.json({ success: true, taskitem });
     } catch(err){
         res.json({ success: false, message: err.message })
@@ -39,7 +41,7 @@ router.delete("/delete/:id", async (req, res) => {
 
 router.get("/:id", async (req, res) => {
     try {
-        const taskitem = await Task.findById(req.params.id);
+        const taskitem = await Task.findOne({_id:req.params.id, createdBy:req.user._id});
         res.json({ success: true, taskitem });
     } catch (err) {
         res.status(500).json({ success: false, message: err.message });
@@ -49,7 +51,7 @@ router.get("/:id", async (req, res) => {
 
 router.put("/update/:id", async (req, res) => {
     try {
-        const taskitem = await Task.findByIdAndUpdate(req.params.id, req.body, { returnDocument: "after" });
+        const taskitem = await Task.findOneAndUpdate({_id:req.params.id, createdBy:req.user._id}, req.body, { new: true });
         res.json({ success: true, taskitem });
     } catch (err) {
         res.json({ success: false, message: err.message });
@@ -64,7 +66,7 @@ router.delete("/delete-multiple", async(req, res)=>{
             return res.json({success:false, message:"no id provided"});
         }
 
-        const result = await Task.deleteMany({_id: {$in: ids}});
+        const result = await Task.deleteMany({_id: {$in: ids}, createdBy:req.user._id});
         res.json({success:true, result});
 
     } catch(err){
